+++ src/main/java/io/github/zeroone3010/yahueapi/HueBridgeDiscoverer.java
@@ -1,0 +1,10 @@
+package io.github.zeroone3010.yahueapi;
+
+import java.util.List;
+
+/**
+ * An interface for an automatic Hue Bridge discovery method.
+ */
+public interface HueBridgeDiscoverer {
+  List<HueBridge> discoverBridges();
+}